refactor(services): PR-3.5 projects/ + presets/ 子包搬迁 — 6 顶层文件 + 2 内部
projects/（顶层 4 文件搬入 services/）：
  studio/projects.py         → services/projects/projects.py
  studio/versions.py         → services/projects/versions.py
  studio/versions_phase.py   → services/projects/phase.py
  studio/project_jobs.py     → services/projects/jobs.py

presets/（services/presets.py + 顶层 presets_io.py 合并到子包）：
  services/presets.py        → services/presets/__init__.py（**直接作为包入口**而非
                               单独 fork.py，因为 services/presets/ 目录已经覆盖了
                               旧 services/presets.py 文件 import 路径，shim 文件无效）
  studio/presets_io.py       → services/presets/io.py

修复发现：
- version_config 跨包 reach into presets_io._absolutize_model_paths（审计 P2 问题 13）
  在此处变成循环 import → 改 from .presets.io import _absolutize_model_paths
- tagging.cltagger / wd14 / presets.__init__ 内部 import model_downloader 走 shim 时
  setattr 监补不传透 → 改用 from ..models import downloader as model_downloader 直接走真实位置

回归：全测试套 1622 passed（vs 上 PR base 1586，新增 36 来自 import smoke 新模块）
